+++ .husky/pre-push
@@ -1,3 +1,4 @@
 #!/usr/bin/env sh
 . "$(dirname "$0")/_/husky.sh"
 
+npm run pre-push
