+++ package.json
@@ -50,6 +50,7 @@
     "husky": "^8.0.3",
     "jsdom": "^22.1.0",
     "lint-staged": "^15.0.2",
+    "msw": "^2.0.8",
     "node-fetch": "^3.3.2",
     "prettier": "^3.0.3",
     "typescript": "^5.0.2",
