+++ src/router.tsx
@@ -4,13 +4,13 @@ import { ItemComponent, itemLoader } from './components/item.component';
 import { ErrorBoundaryComponent } from './components/error-boundary.component';
 import { NotFoundPage } from './components/not-found-page';
 import { Provider } from 'react-redux';
-import { store } from './store/search-value.slice';
+import { setupStore } from './store/store';
 
 export const router = createBrowserRouter([
   {
     path: '/',
     element: (
-      <Provider store={store}>
+      <Provider store={setupStore()}>
         <App />
       </Provider>
     ),
