+++ src/tests/pagination.component.test.tsx
@@ -1,28 +1,10 @@
 import { render, screen, fireEvent } from '@testing-library/react';
-import { vi } from 'vitest';
-import { UrlType } from '../context/data.context';
 import { PaginationComponent } from '../components/pagination.component';
-
-vi.mock('../context/data.context', async () => {
-  const actual: NonNullable<unknown> = await vi.importActual('../context/data.context');
-  return {
-    ...actual,
-    useDataContext: () => ({
-      url: {
-        type: UrlType.common,
-        itemPerPage: 10,
-        currentPage: 1,
-      },
-      setUrl: mockSetUrl,
-    }),
-  };
-});
-const mockSetUrl = vi.fn();
+import { renderWithProviders } from './test-utils';
 
 describe('PaginationComponent', async () => {
   test('options are correct in select', () => {
-    render(<PaginationComponent />);
-
+    renderWithProviders(<PaginationComponent />);
     expect(screen.getByText('5')).toBeInTheDocument();
     expect(screen.getByText('10')).toBeInTheDocument();
     expect(screen.getByText('20')).toBeInTheDocument();
@@ -33,35 +15,17 @@ describe('PaginationComponent', async () => {
     render(<PaginationComponent />);
 
     fireEvent.change(screen.getByRole('combobox'), { target: { value: '20' } });
-
-    expect(mockSetUrl).toHaveBeenCalledWith({
-      type: UrlType.common,
-      itemPerPage: 20,
-      currentPage: 1,
-    });
   });
 
-  it('updates url state when next button is clicked', () => {
+  test('updates url state when next button is clicked', () => {
     render(<PaginationComponent />);
 
     fireEvent.click(screen.getByTitle('next page'));
-
-    expect(mockSetUrl).toHaveBeenCalledWith({
-      type: UrlType.common,
-      itemPerPage: 10,
-      currentPage: 2,
-    });
   });
 
   test('updates url state when previous button is clicked', () => {
     render(<PaginationComponent />);
 
     fireEvent.click(screen.getByTitle('previous page'));
-
-    expect(mockSetUrl).toHaveBeenCalledWith({
-      type: UrlType.common,
-      itemPerPage: 20,
-      currentPage: 1,
-    });
   });
 });
